backend/requirements.txt
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.0
gunicorn==23.0.0
psycopg[binary]==3.3.3
whitenoise==6.11.0
